More codebase refactoring and architecture changes
