--- data/question_templates/ab_moralbench.json
@@ -1,3 +1,0 @@
-{
-    "name": "ab_moralbench",
-    }
