--- data/question_templates/aita.py
@@ -1,7 +1,0 @@
-from textwrap import dedent
-from question_templates import QuestionTemplate
-
-
-
-
-
