--- data/response_templates/ab.json
@@ -1,68 +1,0 @@
-{
-    "responses_A": [
-        "A",
-        "a",
-        "A.",
-        "a.",
-        "A. {optionA}",
-        "A. {optionA_short}",
-        "[A]",
-        "(A)",
-        "**A**",
-        "I would choose A.",
-        "I would choose A",
-        "Answer A",
-        "Answer A.",
-        "Answer: A",
-        "Answer: A.",
-        "Answer A.",
-        "Answer A",
-        "Option A.",
-        "Option A",
-        "Response A",
-        "Response A.",
-        "Response: A",
-        "Choose A",
-        "Choose A.",
-        "I choose option A.",
-        "I choose option A",
-        "Option 1",
-        "{optionA}",
-        "{optionA_short}",
-        "- {optionA}",
-        "- {optionA_short}"
-    ],
-    "responses_B": [
-        "B",
-        "b",
-        "B.",
-        "b.",
-        "B. {optionB}",
-        "B. {optionB_short}",
-        "[B]",
-        "(B)",
-        "**B**",
-        "I would choose B.",
-        "I would choose B",
-        "Answer B",
-        "Answer B.",
-        "Answer: B",
-        "Answer: B.",
-        "Answer B.",
-        "Answer B",
-        "Option B.",
-        "Option B",
-        "Response B",
-        "Response B.",
-        "Response: B",
-        "Choose B",
-        "Choose B.",
-        "I choose option B.",
-        "I choose option B",
-        "Option 2",
-        "{optionB}",
-        "{optionB_short}",
-        "- {optionB}",
-        "- {optionB_short}"
-    ]
-}
