--- data/response_templates/compare.json
@@ -1,57 +1,0 @@
-{
-    "responses_A": [
-        "Yes",
-        "yes",
-        "Yes.",
-        "yes.",
-        "I do",
-        "I do.",
-        "*Yes*",
-        "Answer: Yes",
-        "Answer: Yes.",
-        "answer: yes",
-        "Response: Yes",
-        "Yes, {optionA}",
-        "I prefer {optionA}",
-        "I prefer {optionA_short}",
-        "I prefer \"{optionA}\"",
-        "I prefer \"{optionA_short}\"",
-        "I prefer {optionA} over {optionB}",
-        "I prefer {optionA_short} over {optionB}",
-        "I prefer {optionA_short} over {optionB_short}",
-        "I prefer \"{optionA}\" over \"{optionB}\"",
-        "I prefer \"{optionA_short}\" over \"{optionB}\"",
-        "I prefer \"{optionA_short}\" over \"{optionB_short}\"",
-        "{optionA}",
-        "{optionA_short}",
-        "- {optionA}",
-        "- {optionA_short}"
-    ],
-    "responses_B": [
-        "No",
-        "no",
-        "No.",
-        "no.",
-        "*No*",
-        "*No*",
-        "Answer: No",
-        "Answer: No.",
-        "answer: no",
-        "Response: No",
-        "No, {optionB}",
-        "I prefer {optionB}",
-        "I prefer {optionB_short}",
-        "I prefer \"{optionB}\"",
-        "I prefer \"{optionB_short}\"",
-        "I prefer {optionB} over {optionA}",
-        "I prefer {optionB_short} over {optionA}",
-        "I prefer {optionB_short} over {optionA_short}",
-        "I prefer \"{optionB}\" over \"{optionA}\"",
-        "I prefer \"{optionB_short}\" over \"{optionA}\"",
-        "I prefer \"{optionB_short}\" over \"{optionA_short}\"",
-        "{optionB}",
-        "{optionB_short}",
-        "- {optionB}",
-        "- {optionB_short}"
-    ]
-}
