+++ data/templates/response_templates.py
@@ -1,0 +1,12 @@
+A_TOKENS = ["A", "a", "A.", "a.", "[A]", "(A)", "*A*", "**A**",
+            " A", " a", " A.", " a.", " [A]", " (A)", " *A*", " **A**",
+            "A ", "a ", "A. ", "a. ", "[A] ", "(A) ", "*A* ", "**A** "]
+B_TOKENS = ["B", "b", "B.", "b.", "[B]", "(B)", "*B*", "**B**",
+            " B", " b", " B.", " b.", " [B]", " (B)", " *B*", " **B**",
+            "B ", "b ", "B. ", "b. ", "[B] ", "(B) ", "*B* ", "**B** "]
+YES_TOKENS = ["Yes", "yes", "YES", "Yes.", "yes.", "YES.", "[Yes]", "(Yes)", "*Yes*", "**Yes**",
+              " Yes", " yes", " YES", " Yes.", " yes.", " YES.", " [Yes]", " (Yes)", " *Yes*", " **Yes**",
+              "Yes ", "yes ", "YES ", "Yes. ", "yes. ", "YES. ", "[Yes] ", "(Yes) ", "*Yes* ", "**Yes** "]
+NO_TOKENS = ["No", "no", "NO", "No.", "no.", "NO.", "[No]", "(No)", "*No*", "**No**",
+             " No", " no", " NO", " No.", " no.", " NO.", " [No]", " (No)", " *No*", " **No**",
+             "No ", "no ", "NO ", "No. ", "no. ", "NO. ", "[No] ", "(No) ", "*No* ", "**No** "]
